Enable `c_unwind` feature
The Rust team has put out a call for testing on the `c_unwind` feature
and associated `C-unwind` ABI:

- https://github.com/rust-lang/rfcs/pull/2945#issuecomment-1134003542

I intend to implement mruby's exception unwinding using Rust's panic
infrastructure:

- https://github.com/artichoke/artichoke/issues/35

This commit changes the Rust toolchain to nightly for this test. Merging
this experiment will be blocked on the `C-unwind` ABI making it to
stable.