Disable `exc_throw` in mruby when compiling for Artichoke
When `-DARTICHOKE` is passed when compiling mruby, `#ifndef` out the
definition and implementation of the `exc_throw` static function and
replace it with a forward declaration.

This function normally uses the `MRB_THROW` macro to initiate an unwind.

This prepares us to implement this function in Rust with a
`#[no_mangle] unsafe extern "C-unwind" fn` item that unwinds with a Rust
panic.

The previous commit disables the implementations of `mrb_protect` and
`mrb_ensure` which pair usage of `MRB_TRY` and `MRB_CATCH` with the
`MRB_THROW` in `exc_throw`.