Prototype `mrb_protect` in Rust using `std::panic`
- Add a custom panic payload which wraps a `sys::mrb_value` and
  implements the necessary traits to be used with `panic!`.
- Implement `exc_throw` with `std::panic::panic_any` and the custom
  panic payload.
- Implement `mrb_protect` with `std::panic::catch_unwind` and
  downcasting to check for the custom panic payload.

When using `std::panic::panic_any`, the built-in Rust panic hook is
triggered which prints the panic to stderr and enables the
`RUST_BACKTRACE` handler. This commit includes some code that
experiments with disabling these built-in behaviors.

Per feedback from upstream, `panic_any` should be replaced with
`std::panic::resume_unwind` which initiates the panic with a custom
payload and avoids running the registered panic hook.